Split into gateways; add --outputTransport
