+++ src/types.ts
@@ -1,0 +1,4 @@
+export interface Logger {
+  info: (...args: any[]) => void
+  error: (...args: any[]) => void
+}
